Move usePanel to hooks/ folder
