+++ app/javascript/components/editor/header/ActionMore.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useEffect, useState } from 'react'
 import { Icon } from '../../common/Icon'
 import { BugReportModal } from '../../modals/BugReportModal'
-import { usePanel } from './usePanel'
+import { usePanel } from '../../../hooks/use-panel'
 
 export const ActionMore = ({
   onRevertToLastIteration,
