+++ app/javascript/hooks/use-panel.tsx
