+++ test/javascript/hooks/use-panel.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render, fireEvent, waitFor } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
-import { usePanel } from '../../../../../app/javascript/components/editor/header/usePanel'
+import { usePanel } from '../../../app/javascript/hooks/use-panel'
 
 const TestComponent = () => {
   const {
